+++ spb/backends/base_backend.py
@@ -271,15 +271,14 @@ class Plot:
         # Automatic legend: if more than 1 data series has been provided
         # and the user has not set legend=False, then show the legend for
         # better clarity.
-        self.legend = _legend = kwargs.get("legend", None)
-        if not self.legend:
-            self.legend = False
-            if (len([s for s in self._series if s.show_in_legend]) > 1) or (
-                any(s.is_parametric and s.use_cm for s in self._series)):
+        self.legend = kwargs.get("legend", None)
+        if self.legend is None:
+            # if len([s for s in self._series if s.show_in_legend]) > 1:
+            if len([s for s in self._series
+                if s.show_in_legend and (not s.use_cm)]) > 1:
                 # don't show the legend if `plot_piecewise` created this
                 # backend
-                if ((not ("process_piecewise" in kwargs.keys())) and
-                    ((_legend is True) or (_legend is None))):
+                if not ("process_piecewise" in kwargs.keys()):
                     self.legend = True
 
         # Objects used to render/display the plots, which depends on the
@@ -479,12 +478,6 @@ class Plot:
         series.extend(self.series)
         series.extend(other.series)
         kwargs = self._do_sum_kwargs(self, other)
-        # If the first plot (`p1`) of the summation has been created without
-        # specifying `legend`, then `p1.legend` might be False, hence
-        # `kwargs["legend"]` might be False. But when adding multiple plots
-        # it is very likely that user expect a legend to be shown. Hence,
-        # reset legend and let the backend decide if it needs one or not.
-        kwargs["legend"] = None
         return type(self)(*series, **kwargs)
 
     def append(self, arg):