+++ spb/ccomplex/complex.py
@@ -402,6 +402,10 @@ def plot_real_imag(*args, **kwargs):
     backend : Plot, optional
         A subclass of ``Plot``, which will perform the rendering.
         Default to ``MatplotlibBackend``.
+    
+    colorbar : boolean, optional
+        Show/hide the colorbar. Only works when ``use_cm=True`` and 3D plots.
+        Default to True (colorbar is visible).
 
     detect_poles : boolean, optional
         Chose whether to detect and correctly plot poles. Defaulto to False.
@@ -787,6 +791,9 @@ def plot_complex(*args, **kwargs):
         Specify the colormap to be used on enhanced domain coloring plots
         (both images and 3d plots). Default to ``"hsv"``. Can be any colormap
         from matplotlib or colorcet.
+    
+    colorbar : boolean, optional
+        Show/hide the colorbar. Default to True (colorbar is visible).
 
     coloring : str or callable, optional
         Choose between different domain coloring options. Default to ``"a"``.
@@ -1415,6 +1422,9 @@ def plot_complex_vector(*args, **kwargs):
     backend : Plot, optional
         A subclass of `Plot`, which will perform the rendering.
         Default to `MatplotlibBackend`.
+    
+    colorbar : boolean, optional
+        Show/hide the colorbar. Default to True (colorbar is visible).
 
     contours_kw : dict
         A dictionary of keywords/values which is passed to the backend
@@ -1860,6 +1870,7 @@ def plot_riemann_sphere(*args, **kwargs):
     if kwargs.get("threed", False):
         kwargs.setdefault("xlabel", "Re")
         kwargs.setdefault("ylabel", "Im")
+        kwargs.setdefault("zlabel", "")
         colorbar = kwargs.pop("colorbar", True)
         Backend = kwargs.get("backend", THREE_D_B)
         t, p = symbols("theta phi")