+++ spb/vectors.py
@@ -337,6 +337,9 @@ def plot_vector(*args, **kwargs):
     backend : Plot, optional
         A subclass of `Plot`, which will perform the rendering.
         Default to `MatplotlibBackend`.
+    
+    colorbar : boolean, optional
+        Show/hide the colorbar. Default to True (colorbar is visible).
 
     color_func : callable, optional
         Define the quiver/streamlines color mapping when ``use_cm=True``.